+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added Abundant Incense mythic ability
 * Added Abundant Lay On Hands mythic ability
 * Added Abundant Fervor mythic ability
+* Added Second Patron mythic ability
 
 ## Version 1.10.4
 * Now works with 1.1.6e